Add common gRPC configuration settings
* Change opencensusexporter gRPC configuration settings to the new common configuration
